% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernal.R
\name{k_qdap}
\alias{k_qdap}
\title{k_qdap}
\usage{
k_qdap(x, y, xnew, lambda = 0, iter = 1, method = "Penalization")
}
\description{
This function executes the QDA by Projection method with kernel trick
}
\details{
Place holder
}
