% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{maxquadratio}
\alias{maxquadratio}
\title{Maximizing the ratio of two full rank quadratic forms}
\usage{
maxquadratio(sigma0, sigma1)
}
\arguments{
\item{sigma0}{The matrix which determines the top full rank quadratic form}

\item{sigma1}{The matrix which determines the bottom full rank quadratic form}
}
\value{

}
\description{
Maximizing the ratio of two full rank quadratic forms
}
\details{
Find the direction `a' that maximizes (t(a) \%\emph{\% sigma0 \%}\% a)/(t(a) \%\emph{\% sigma1 \%}\% a)
}
\author{
Ruiyang Wu
}
