% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdap.R
\name{qdap}
\alias{qdap}
\title{qdap}
\usage{
qdap(x, y, xnew, lambda = 0, iter = 1, method = "Penalization")
}
\arguments{
\item{x}{training predictors}

\item{y}{training labels}

\item{xnew}{test predictors}

\item{lambda}{}

\item{iter}{number of iterations}

\item{method}{}
}
\value{
a list of the predicted labels, optimal direction and convergence status
}
\description{
This function executes the QDA by Projection method
}
\details{
Place holder
}
\author{
Ruiyang Wu
}
