% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis_rate.R
\name{mis_rate}
\alias{mis_rate}
\title{mis_rate function}
\usage{
mis_rate(a, mu0, mu1, sigma0, sigma1)
}
\arguments{
\item{a}{direction}

\item{mu0}{mean of class 0}

\item{mu1}{mean of class 1}

\item{sigma0}{covariance matrix of class 0}

\item{sigma1}{covariance matrix of class 1}
}
\value{
classification error
}
\description{
This is the classification error/misclassification error function
}
\details{
Place holder
}
\author{
Ruiyang Wu
}
