% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdap.R
\name{qdap}
\alias{qdap}
\title{Qdadratic Discriminant Analysis by Projection.}
\usage{
qdap(
  x,
  y,
  xnew = NULL,
  lambda = 0,
  iter = 1,
  method = "Penalization",
  optim = "codesc"
)
}
\arguments{
\item{x}{A matrix containing the predictors of the training data.}

\item{y}{A vector containing the class labels of the training data.}

\item{xnew}{A matrix containing the predictors of the test data.}

\item{lambda}{The tuning variable used for either the "Penalization" method or the "Thresholding" method.}

\item{iter}{Number of iterations to apply QDAP. If greater than 1, this will keep searching the optimal direction in the orthogonal complement of the previous optimal subspace. (Beta)}

\item{method}{A method to get sparse optimal direction. "Penalization" for penalizing over the l2 norm of the optimal direction, "Thresholding" for thresholding over each entry of the optimal direction. (Beta)}

\item{optim}{The optimization method used for the misclassification function. "BFGS" for Broyden–Fletcher–Goldfarb–Shanno algorithm, "codesc" for coordinate descent algorithm.}
}
\value{
a list of the predicted labels, optimal direction and convergence status
}
\description{
This function runs the Quadratic Discriminant Analysis by Projection (QDAP) method.
}
\details{
Place holder
}
\author{
Ruiyang Wu
}
